# lib/scanner: Reduce allocations in Blocks