Here's a small optimization for scanner.Blocks. It performs fewer allocations and gets a tiny, but consistent, throughput boost:

```
name        old time/op    new time/op    delta
HashFile-8    88.3ms ± 1%    87.8ms ± 0%   -0.58%  (p=0.000 n=19+19)

name        old alloc/op   new alloc/op   delta
HashFile-8    45.4kB ± 0%    45.3kB ± 0%   -0.21%  (p=0.000 n=20+18)

name        old allocs/op  new allocs/op  delta
HashFile-8      28.0 ± 0%      24.0 ± 0%  -14.29%  (p=0.000 n=20+18)

name        old speed      new speed      delta
HashFile-8   202MB/s ± 1%   203MB/s ± 0%   +0.58%  (p=0.000 n=19+19)
```