@calmh 2020-06-01
More generally this is something performing file I/O and hashing, I doubt the four allocations removed have any impact on the overall performance. (Cleanup can still be good if the end result is clearer or less code...)