@greatroar 2020-06-05
I can't claim this is a cleanup since the current code is pretty straightforward. Closing; the trouble of changing the code is not worth it.